log to stderr
