+++ src/babeld.c
@@ -34,7 +34,7 @@ void babeld_parse_line(struct context *ctx, char *line) {
 			free(ifname);
 			n = sscanf(line, "%ms neighbour %*s", &action);
 			if (n == 2 ) {
-				printf("Received changes on a neighbour but could not match them on any of the neighbour-patterns. Exiting parser %d. This is a bug that should be reported.\n", n);
+				log_verbose(ctx, "Received changes on a neighbour but could not match them on any of the neighbour-patterns. Exiting parser %d. This is a bug that should be reported.\n", n);
 			}
 
 			goto end;
@@ -43,7 +43,7 @@ void babeld_parse_line(struct context *ctx, char *line) {
 	struct in6_addr address;
 
 	if (inet_pton(AF_INET6, address_str, &address) != 1) {
-		printf("received garbage instead of IPv6-address, not parsing line from babeld\n");
+		log_verbose(ctx, "received garbage instead of IPv6-address, not parsing line from babeld\n");
 		goto end;
 	}
 
@@ -88,7 +88,7 @@ bool babeld_handle_in(struct context *ctx, int fd) {
 		line = strsep(&stringp, "\n");
 
 		if ((ctx->verbose) && (strlen(line) > 1))
-			printf("about to parse line: %s\n", line);
+			log_verbose(ctx, "about to parse line: %s\n", line);
 
 		if (stringp == NULL)
 			break; // no line found
