+++ src/mmfd.h
@@ -1,7 +1,9 @@
 #pragma once
 
 #include "vector.h"
+#include "error.h"
 
+#include <stdio.h>
 #include <stdint.h>
 #include <stdbool.h>
 #include <netinet/in.h>
@@ -30,3 +32,5 @@ struct neighbour {
 	int reach;
 	int cost;
 };
+
+void log_verbose(struct context *ctx, const char *format, ...);
